+++ server/api/file/index.unit.test.js
@@ -1,0 +1,8 @@
+/* globals describe, expect, test */
+import router from './index';
+
+describe('File API Route Handlers:', () =>
+  test('should have unchanged route handlers', () =>
+    expect(router.stack).toMatchSnapshot()
+  )
+);