+++ server/routes.js
@@ -9,6 +9,7 @@ import HistoryApi from './api/history';
 import ProductApi from './api/product';
 import ClassApi from './api/class';
 import LocationApi from './api/location';
+import FileApi from './api/file';
 import Auth from './auth';
 
 export default app => {
@@ -21,6 +22,7 @@ export default app => {
   app.use('/api/product', ProductApi);
   app.use('/api/class', ClassApi);
   app.use('/api/location', LocationApi);
+  app.use('/api/file', FileApi);
   app.use('/auth', Auth);
 
   // Return Apple's merchant ID domain association file